Break out components from page
